+++ src/main/java/cc/polyfrost/oneconfig/gui/elements/config/ConfigHeader.java
@@ -29,7 +29,7 @@ public class ConfigHeader extends BasicOption {
     }
 
     @Override
-    public int getHeight() {
+    public float getHeight() {
         return 32;
     }
 }
